Adding the tests for the base monitoring
This shifts some of the tests from base tests to stateful tests, where
files are added, changed, and removed, and the model checks that between
any two scans, the data discovered on disk matches what the operation
logically does when the disk is abstracted.

A stateless test is added for the private function doing the diff
between two sets used pre-formatted deterministic runs.

The data set across runs should therefore be trustworthy.