Adding event sending from scans
A polling process can be started, which will poll directories at
predefined intervals. All the changes detected will be exported as
events.

Currently the events are exported as batches and it's not _that_ useful,
but it will make sense when inotify and other OS-specific interfaces are
allowed to do streaming updates.

Events are registered through gproc. A property suite has been added.