Add foundations of a directory model
model-based approach: create a tree/forest data structure, where each
inner node is a directory, and each leaf is a file; at any given level,
a file can't bear the same name as an inner node.  Files are created or
managed by picking a random path in the tree and inserting data in it.
The tree represents the state and can be used to do all operations, but
without caring about the filesystem.

The advantage of the tree model is that it can be used for both
stateless and stateful tests. Additional inner node types can be added
to represent things such as symlinks (where supported), or used to
create alterations such as "case sensitive duplicates" of given paths.