Add a connection step
This prepares the ground for TCP handling by having specific steps to
set up a connection with a peer, and to remove them as well. These steps
are then used within the sync and id operations, which are
client->server and peer-specific, demanding that a connection is both
set up and torn down before and after the operation.

This also starts to show the limits of the current FSM format, and may
demand switching towards a client vs. server FSM for syncing and using
enter_loop to deal with a bunch of things.