Reworked the FSM
Now works through using way more explicit states, rather than the
callback mode with a lot of interleaving.

I figure the lesson here is to really counter-balance whether there is a
set of limited client/server-style interactions between a caller and the
FSM, which can work well with handle_event_functions, or whether there
are a lot of potential states where specific client calls are only
permissible in a subset of them and where the data handling can vary
significantly.

Because in this case, revault_fsm is a lot more legible than
revault_sync_fsm was, even with the introduction of intermediary states
used only for transitions like connecting/disconnect/initialized.

I can also imagine benefits in later observability and tracing
mechanisms.