Don't accidentally initialize a FSM
This tiny typo would cause crashes, as invariants in the initialized
state enter checks would detect an undefined ID in an initialized state
and would die.

This is of course because a read-only check to an ID shouldn't turn the
FSM into an internal initialized mode without going through actual
initialization.