Fix maestro loader, auto-boot FSMs from config
Test highlights an issue in TLS validation where the tests for
revault_fsm rely on local names to make things work but that name
doubles as the directory name being monitored and won't fit with the
purposes of messaging _and_ in-band validation.

So the callback now allows overriding that on demand to make the
internal tests work, and holds true for general defaults.

With this, the maestro loader initialization shows an end-to-end
exchange over TLS using two distinct nodes.