Add a per-lineage UUID check
Each server once booting for the first time sets up a UUID. This UUID is
passed around along with the ID upon a fork operation, such that all
peers that are part of a network share the same UUID.

This is a minimal check done freely by the client (which is trusted via
certs in the case of TLS, but is totally unauthentified in the case of
TCP and disterl) which sends in its own UUID to the server when
establishing a connection.

The server checks that UUID against its own, and if it notices a clash,
it will bail out of the connection and refuse seeing it established.

The main objective of this is to prevent the accidental case where you
misconfigure a client and make it talk to an unintended server and then
accidentally start corrupting stuff by having broken ID ranges and
unexpected files moving around.

This is _not_ a security feature because the trust model here would let
a peer just send an unchecked message (by omitting the UUID, since
that's a thing needed for the first connection) and bypass the whole
thing anyway.

If we wanted to grow the threat model there, we'd need to divide the
state machine of the server's side into two new halves (UUID submitted
vs. no UUID submitted) and restrict operations that can be done
unauthentified. This isn't the expected model here however and we don't
cover that element since it assumes you've leaked your keys already and
welp, not much you can do to protect against damage once this is done.