attempt a different buffering scheme for refc bins
After looking for the long tail of slow transfers and finding
out that adding `{active,N}` batch buffering didn't help much, as a packet
capture still showed filled TCP window.

I noted that the process memory never increased (due to using large refc
binaries which go into progressively larger single block carrier memory
pools I assume) and the number of reductions slowed down, I suspect the
issue could be the allocation of single block carriers as we keep
appending binaries and forcing reallocations of the whole thing, mostly
since the process would rarely hit a minor GC even if it had to store
megabytes of content, and so more work was done just dealing with the
binary appends and progressively slowed down buffer consuming from the
socket.

So the alternative buffering mechanism here keeps all binaries received
into a list and sums up their sizes, and only merges them as one big
binary to deserialize once the required size is reached. The theory is
that this lets us store most refc binaries in multiblock carriers until
the final merge which would be a single reallocation into a single block
carrier for deserialization, and makes sure we rarely if ever hit slower
allocators.

Measurements show that the same file transfer is now taking a fraction
of the old one; the slowest receive which was at 31ms is now at 0.645ms,
and the whole transfer mechanism is made ~20x faster (~20 seconds to
right below 1 second for a 28MB file on local wifi) by this small scheme,
but the bigger the file, the more notable the speed gain.