Reduce packet recv instrumentation
Sending one span per recv calls makes the traces noisy and nearly
useless, particularly during large file transfers.

Instead, this patch reduces the number of spans to one per message batch
where messages are successfully unpacked. This will allow to create a
sort of aggregate value where the number of bytes in the buffer can be
compared to the duration of the span to derive transfer rates, and I've
also added a trailing value to better track what was received and
consumed vs. received and left for the next batch.

This should bring back the traces to a manageable level now that
performance woes are mostly taken care of, or at least brought to a
reasonable level.