Support config reload, client/server mode switch
Support reloading configuration reloading to alter running state of
ReVault. This is currently done with a hard stop and restart of all
trackers, FSMs, and servers and clients. This is not elegant, but it is
sufficient to support simpler tests for mode switch, which is the core
feature implemented here.

Mode switching between client and server mode is now supported once the
FSM is initialized (an ID and UUID is chosen). The mode switch is
handled by going back to initialized mode and handing off the state
transition with a postpone.

It will only work if a FSM has both configurations, and the mode switch
can be bypassed by straight up calling for a sync or receiving a
mode-triggering message (connection, sync) in either the client or
server mode, but not while a sync is ongoing.

This will allow to fully support de-centralized topologies where many
servers can exist and sync between each other depending on who they talk
to.