Weave in ignore list of regexes all through app
This is a lot more work than I thought, because this is a lot more stuff
than expected. But this passes the ignore list through all layers that
require them -- mostly the dirmon tracker and the dirmon poll
operations, but since these are part of a whole instantiation hierarchy
starting with maestro and going through the FSM, well here we are.

The default ignore list now includes .DS_Store files, but this isn't
tested yet. This commit only brings parity when the empty ignore list is
passed; more tests are required to demonstrate support of the new change
(and of how it can migrate away from the old one via the tracker
filtering out its snapshot on load).