Support Ignoring files (#34)
This is a lot more tedious work than I thought, because this is weaving
the list of regexes through a lot more stuff than expected -- mostly the
dirmon tracker and the dirmon poll processes that need it directly,
but since these are part of a whole instantiation hierarchy
starting with maestro and going through the FSM, well here we are.

The default ignore list now includes .DS_Store files. Tests demonstrate
support of the new change (and of how it can migrate away from
an old ignore list via the tracker filtering out its snapshot on load).