Extract hash scan to be s3-friendly
S3 Standard costs:

- $0.005 for PUT, COPY, POST and LIST requests
- $0.0004 for GET, SELECT, and all other requests.

Gotchas:

- LIST can't return the checksums
- LIST will require to be split in multiple paginated calls after ~3000
  elements
- easily getting the checksums can be done with a HEAD request with
  args, so long as said checksum was submitted
- rename doesn't exist but can be approximated with COPY+DELETE

The scan operation we do is likely costliest because it will be repeated
every time we sync and in line with the directory size.

This implies doing:

- a list of the directory ($0.005)
- for each file a HEAD request to get the hash (100*$0.0004=$0.04)

What we could however do is cache the hashes and pair them with the item
name and a lastmodified stamp which can be extracted at the same time as
we get filenames from the list operation. Using the last modified times,
we can know when to re-compute hashes, which lets us re-fetch only changed
hashes or new files (which should have no entry). This cache itself
could be maintained in a file stored on S3 in a known location (eg.
alongside the directory tracked state) for cheaper than the repeated
HEAD queries.

This is a worthy approach that could cut costs almost tenfold (for 100
files at least; it's less for fewer files and more for more files), but
requires moving file abstractions a different way for the scan, which is
part of the core tracking mechanism.

So this commit does just that by splitting the file iteration and
filtering/sorting logic between the dirmon and file modules, which in
turn removes the need for the filelib:fold_files/5 wrapper, which
becomes an implementation detail.