Make a basic S3 sync work end-to-end
There's more stuff to test tagged in TODOs

Over bigger directories, the bigger problem is that the data flow is
fully asynchronous when sending from the fsm to the tls server handler,
and from the tls client to the fsm.

This is generally not a problem when syncing because in normal mode, the
local disk tends to be faster than the network over large file
transfers. However, in S3 mode as I tested it (with two nodes on the
same host), then suddenly the slowest link is now the network between
the s3-handling node and aws s3 itself.

   [disk] --> [local A] ---ext network----> [local B] --> [disk]
                              |
                              '- bottleneck

   [disk] --> [local A] ---local network--> [local B] --> [s3]
                                                       |
                                           bottleneck -'

This appears to lead to all the reading from local disk to be done at
once and transfered over the loopback interface to the s3-handling node,
which bubbles it all up in memory until it sort of freezes over and the
node becomes unresponsive.

It's gonna be time to start doing flow control better.