Handle s3 client cache refresh in system time
If the system time skips around (eg. the computer goes to sleep then
comes back up, or the DST switches) then the timer we schedule based on
the monotonic clock isn't going to work since it will wait too many
ticks to trigger.

We keep that check (it's good to be proactive when possible) but then
also do an extra check based on the system time before handing off a
client. If we'd time out, refresh the client.

This also comes with a retry limit that ensures a cooldown period and
avoids infinite loops.