Ignore deleted file in diff seeding
The files scan mechanism is meant to provide the equivalent of a
previous directory snapshot. However, files that were deleted (and
tombstoned) would show up as if they existed, which forced to do a bunch
of pre-deletions on each boot.

This wasn't too costly on a local disk, but it represented a significant
delay on boot time on S3-backed mechanisms since each snapshot change
may result in a file being sync'd. After a while, hundreds of deletes
were detected on first boot in a way that wasn't necessary.