Test all S3 modules; move is_file to is_regular
As it turns out, S3 does not allow is_regular-like usage and calling
HEAD on a directory isn't guaranteed to return a result. That's alright
though because the use that was being made by the tracker behind it all
only actually cares about regular files in the first place, so change
the whole working set for is_regular.

All the other significant calls are mocked out based on manually
verified S3 headers and bodies being returned, so we can get ongoing
tests of s3 operations without doing the full integration suite that
costs money.