Adding cross-compilation profile
The cross-compilation profile works on a simple basis:

- the target system where the code will run has a pre-built Erlang
  runtime installed with SSL/TLS support (eg. `sudo apt-get install
  erlang-nox` for a wx-free Erlang on a debian or ubuntu host) only
- the host where the build is made has a compatible erlang version used
- the build host calls `rebar3 as portable do release, tar`
- the tarball at `_build/portable/rel/revault/revault-$VSN.tar.gz` is
  moved to the target host
- the target host unpacks the tarball with `tar -xvf revault-$VSN.tar.gz`
- the target host gets the proper TOML configuration and file layout
- the target host calls ./bin/revault console

The target host then runs the software and is possible to contact for
sync jobs.

This is not the same as a fully portable build that would be
self-contained, but it is much simpler to put in place than having to
set up a whole build environment for each target system, and will work
so long as none of the non-system libraries used require NIFs.