Rework file transfer queuing mechanism
The old queuing mechanism mostly worked by putting all the message in
the `gen_statem`'s `next_event` queue. This is not a great mechanism
because it it has some limitations:

> The stored events are inserted in the queue as the next to process
> before any already queued events. The order of these stored events is
> preserved, so the first next_event in the containing list becomes the
> first to process.

The issue here is that every time we send ourselves a `ping` message,
for example, it will go the back of the whole queue. In some cases I
suspect this could lead to issues in flow control (eg. the TLS server
waits for a `pong` to send messages, but the FSM is stuck trying to send
more data).

Instead, we start using a queue data structure to more literally hold
the messages that are to be processed later.

This may also become useful if we try to support sending large files by
breaking them in sub-messages (eg. 15mb per batch) which would just pop
them in front of the FSM's queue. Now with a data structure that's more
explicit, we'll have the ability to control the scheduling without
preventing other messages from being handled in between.