Prepare data wrapper change by bumping version
The data wrapper change introduces a new operation, which means we need
to bump the internal version. At this point in time, the significance of
this is that we eventually need to test the failures of reading the
proper messages, but first let's change the version.

This instantly highlighted shared concerns across _a lot_ of files for
all wrappers, and some inconsistencies, so these get fixed up and
brought under a better centralized umbrella.

Existing tests pass, showing that we're at least consistent within a
single version.