Refactor FSM file sending and scheduling
Pick the comon aspects to the client and the server, where files are
read and broken up to be sent, and unite them within a shared structure.

This further splits up the handling of scheduling, serializing, and the
callback management.

It also temporarily undoes tracing support, which will need to be
reintroduced later.