Filter out directories from listing
If you start a directory monitor as a server in S3 mode and do so by
creating a directory object, it will get returned in listing queries.
Unfortunately, the list calls will require an `is_regular/1` check
(which would be "costly") to filter out.

Instead, add an implicit check in S3 where an object with a trailing `/`
is considered to be a directory and is ignored. This is generally safe
since MacOS and Linux forbid `/` in paths and shouldn't happen outside
of such manual S3 object creations.