Fix deletion conflict issues
Because of how conflicts are transmitted file by file, there's always
potential for a delete conflict to happen, where 2 or more hosts have
concurrently deleted a file without synchronization.

We could _merge_ both deletes and bump the version, but this could run
the risk of ignoring trailing conflicts in a transfer. Instead, we go
explicit by adding deletion conflict messages to the protocol, and
creating empty conflict files when such conflicts are detected. The user
then has to delete the conflict file to resolve everything.